전광찬 연동  서버 연동 부제 운행 제한 오류 수정

+++ src/main/java/com/bnids/gateway/dto/SignageServerRequestDto.java
@@ -34,5 +34,5 @@ import lombok.Setter;
 public class SignageServerRequestDto {
     private String carNo;
     private Long gateId;
-    private Integer registItemId;
+    private Long registItemId;
 }
